+++ CHANGELOG.md
@@ -5,9 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.7 - 2022-08-08
+
+### Added
+
+* plugin settings for ignoring some issue attributes when synchronising
+* git hook scripts running automatically rubocop, brakeman, and tests
+
+### Fixed
+
+* requirement of optional settings to be set
+
 ## 0.1.6 - 2022-07-08
 
-## Fixed
+### Fixed
 
 * nil error during object generation when module issue sync is not fully configured
 
@@ -28,7 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deleted
 
-* blank line from select_tag in sync_issues/_settings.html.erb
+* blank line from select_tag in issue_sync/_settings.html.erb
 
 ### Added